# feat(widgets): add battery-optimized timer widget with gesture controls