- Implement widtimer widget following BangleJS naming conventions
- Add double-swipe protection against accidental activation
- Support gesture controls: swipe right/left ±1min, up/down ±10min
- Include battery optimization with adaptive refresh rates (10s/1s)
- Provide visual feedback with color-coded states
- Generate 3-pulse completion notification over 5 seconds
- Use vector font for crisp display
- Store persistent timer state across device restarts

The widget requires double-swipe in same direction to unlock controls, then allows single swipes for 10 seconds before auto-locking. Refresh rate automatically switches from 10s to 1s in final minute.

🤖 Generated with [Claude Code](https://claude.ai/code)